✨ Refactor assets into one class
